알림에서 상세페이지 바로가기 리뷰 중복 불가 신고 기능 수정

+++ src/main/java/com/goottflix/movie/controller/MovieController.java
@@ -76,7 +76,7 @@ public class MovieController {
     @PostMapping("/movie/review")
     public String reviewPost(@RequestParam("movieId") Long movieId,
                              @RequestParam("userId") Long userId,
-                             @RequestParam("rating") Long rating,
+                             @RequestParam("rating") int rating,
                              @RequestParam(name = "review",required = false) String review){
 
         Review review1 = new Review();
